JPEG2000_CustomDecoder/requirements.txt
# Dependencies for xp-python scripts

# Core
numpy>=1.24.0

# Scene detection
scenedetect>=0.6.0

# Image handling (Pillow is lighter than OpenCV)
Pillow>=10.0.0

# TIFF support for float32/16-bit images
tifffile>=2023.0.0

# Optional: OpenCV for advanced image processing
# opencv-python>=4.8.0
